sum of digits of a number using recursion
In this C program, we are reading the integer number using the ‘num’ variable. The sum() method is used to check the value of ‘num’ variable is not equal to 0. If the condition is true, then execute the statement. Compute the division of the value of ‘num’ variable by 10 integer value. Add the resulted value along with the modulus of the value of ‘num’ variable and return the value to ‘result’ variable. Print the sum of digits of a number.